Nix ruby 1.8.7 support formally
It didn't really exist previously either.  Simplecov was required
unilaterally in spec_helper and I18n was invoked in processable.rb.
Neither of those exists for 1.8.7, so it is unclear when the last time
builds could have passed was.  Ruby 1.8.7 extended support ended
2014-07-31 and it has been end-of-life since, so supporting it further
is inadvisable.  In fact, we are trying to phase out 1.9.3 too, but this
is a start.

Applications: get with the times (2.x) or pin to an older version of
this gem.