to_solr must modify *and* return the passed in hash
The reasons are legacy and strange, but that's how it is.  That means,
you cannot arbitrarily do assignment to the passed in hash and expect it
to work.  You must *modify* the *existing* hash, so that means `.merge!`
instead of `.merge`

In 5.x this whole concern is toast, but this shouldn't block our 4.x
builds from passing.