Fix outdated call to #next_pid in SuriService
If `Config.suri.mint_ids` is false, the SuriService raises a NoMethodError calling `#next_pid` on a Rubydora::Repository instance. Rubydora::Repository instances do not define this method; the method is defined on Rubydora::RestApiClient which can be accessed via `#api` on the repository. Additionally, the xpath that is used in the same `else` clause no longer parses pids out of the `pidList`. I have added code that tries a second XPath if the first fails.

I made these changes so the latest hydra_etd work can use dor-services without patching around the SuriService.