Split the indexing out of the model concerns
This refactor helps to clarify the separate roles of the models and the
indexers. This helps do away with the very deep inheritance structure
that was previously utilized by to_solr in favor of separate classes.
The intention of this PR was purely to improve maintainability and
readability and should not have changed any behavior or added any
features.